Fix Diff2 shift, push a  and push b operations
Before this commit, they were implemented to walk from the top-left of the matrix.
This could cause it to follow paths that would not lead to the optimum solution.
This commit makes those operations walk from the bottom-right of the matrix.
Starting from the bottom-right we make sure the optimum solution is chosen.