Set EMPTY constant to None
Using special characters to represent the empty space is not a good option because they are not printable or already have a well-defined behavior in the print function.
Using non-printable characters is neither a good idea because we cannot see them, so debugging gets harder.
The None value seems the best choice.