Implement TreeDiff2._treedist method (possible bug)
The resulting 6x6 forest distance table is different from the paper* where I got this algorithm from.
In the paper, the 6x6 table is
t6_6 = [[0, 1, 2, 3, 4, 5, 6],
        [1, 0, 1, 2, 3, 4, 5],
        [2, 1, 0, 1, 2, 3, 4],
        [3, 2, 1, 2, 3, 4, 5],
        [4, 3, 2, 1, 2, 3, 4],
        [5, 4, 3, 2, 3, 2, 3],
        [6, 5, 4, 3, 3, 3, 2]]
(lines 3, 4, and 5 changed, starting at position [3,3]).
Apparently that is because it is not allowed to replace node c by node d.
Gotta find out why.

* ZHANG, K. and SHASHA, D. Simple Fast Algorithms For the Editing Distance Between Trees and Related Problems. 1989. Available at: https://grantjenks.com/wiki/_media/ideas/simple_fast_algorithms_for_the_editing_distance_between_tree_and_related_problems.pdf